+++ packages/hydrogen/src/components/Seo/tests/DefaultPageSeo.test.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import {mount} from '@shopify/react-testing';
-import {Helmet} from 'react-helmet-async';
+import {Helmet} from '../../../client';
 
 import {DefaultPageSeo} from '../DefaultPageSeo.client';
 import {TitleSeo} from '../TitleSeo.client';
 import {DescriptionSeo} from '../DescriptionSeo.client';
 import {TwitterSeo} from '../TwitterSeo.client';
 
-jest.mock('react-helmet-async', () => ({
+jest.mock('../../../client', () => ({
   Helmet({children}) {
     return children;
   },