+++ packages/hydrogen/src/foundation/ShopifyProvider/index.ts
@@ -1,1 +1,2 @@
-export {ShopifyProvider} from './ShopifyProvider';
+export {ShopifyProvider} from './ShopifyProvider.server';
+export {ShopifyContext} from './ShopifyProvider.client';