+++ packages/hydrogen/src/foundation/useShop/use-shop.tsx
@@ -1,34 +1,22 @@
 import {useContext} from 'react';
-import {
-  ShopifyContext,
-  makeShopifyContext,
-} from '../ShopifyProvider/ShopifyContext';
-import type {ShopifyContextValue} from '../ShopifyProvider/types';
-import type {ShopifyConfig} from '../../types';
+import {ShopifyContext} from '../ShopifyProvider';
+import {useServerRequest} from '../ServerRequestProvider';
+// @ts-ignore
+import {META_ENV_SSR} from '../../utilities/meta-env-ssr';
 
-let contextValue: ShopifyContextValue | null = null;
+// let contextValue: ShopifyContextValue | null = null;
 
 /**
  * The `useShop` hook provides access to values within `shopify.config.js`. It must be a descendent of a `ShopifyProvider` component.
  */
-export function useShop(): ShopifyContextValue {
-  /**
-   * During RSC, context is not (yet) allowed on Server Component, so we should be calling `setShop`
-   * in server and returning the object here.
-   */
-  if (contextValue) {
-    return contextValue;
-  }
+export function useShop() {
+  const config = META_ENV_SSR
+    ? useServerRequest().ctx.shopifyConfig
+    : useContext(ShopifyContext);
 
-  const context = useContext(ShopifyContext);
-
-  if (!context) {
+  if (!config) {
     throw new Error('No Shopify Context found');
   }
 
-  return context;
-}
-
-export function setShop(shopifyConfig: ShopifyConfig) {
-  contextValue = makeShopifyContext(shopifyConfig);
+  return config;
 }