+++ packages/hydrogen/src/framework/Hydration/ServerComponentRequest.server.ts
@@ -1,4 +1,7 @@
+import type {ShopifyContextValue} from '../../foundation/ShopifyProvider/types';
 import {getTime} from '../../utilities/timing';
+import {HelmetData} from 'react-helmet-async';
+import type {RealHelmetData} from '../../foundation/Helmet/Helmet';
 
 let reqCounter = 0; // For debugging
 const generateId =
@@ -21,7 +24,12 @@ export class ServerComponentRequest extends Request {
   public id: string;
   public time: number;
   // CFW Request has a reserved 'context' property, use 'ctx' instead.
-  public ctx: {cache: Map<string, any>; [key: string]: any};
+  public ctx: {
+    cache: Map<string, any>;
+    helmet: RealHelmetData;
+    shopifyConfig?: ShopifyContextValue;
+    [key: string]: any;
+  };
 
   constructor(input: any);
   constructor(input: RequestInfo, init?: RequestInit);
@@ -42,7 +50,10 @@ export class ServerComponentRequest extends Request {
     this.time = getTime();
     this.id = generateId();
 
-    this.ctx = {cache: new Map()};
+    this.ctx = {
+      cache: new Map(),
+      helmet: new HelmetData({}) as unknown as RealHelmetData,
+    };
     this.cookies = this.parseCookies();
   }