+++ packages/playground/server-components-worker/src/App.server.jsx
@@ -1,14 +1,18 @@
 import {DefaultRoutes} from '@shopify/hydrogen';
+import {ShopifyProvider} from '@shopify/hydrogen';
 import {Suspense} from 'react';
+import shopifyConfig from '../shopify.config';
 
 export default function App({...serverState}) {
   return (
     <Suspense fallback={'Loading...'}>
-      <DefaultRoutes
-        pages={serverState.pages}
-        serverState={serverState}
-        fallback="Not Found"
-      />
+      <ShopifyProvider shopifyConfig={shopifyConfig}>
+        <DefaultRoutes
+          pages={serverState.pages}
+          serverState={serverState}
+          fallback="Not Found"
+        />
+      </ShopifyProvider>
     </Suspense>
   );
 }